+++ src/dream/QueryDriver/Kysely.ts
@@ -17,6 +17,7 @@ import pluralize from 'pluralize-esm'
 import writeSyncFile from '../../bin/helpers/sync.js'
 import { CliFileWriter } from '../../cli/CliFileWriter.js'
 import DreamCLI from '../../cli/index.js'
+import { pgErrorType } from '../../db/errors.js'
 import _db from '../../db/index.js'
 import associationToGetterSetterProp from '../../decorators/field/association/associationToGetterSetterProp.js'
 import PackageManager from '../../dream-app/helpers/PackageManager.js'
@@ -32,6 +33,7 @@ import MissingThroughAssociationSource from '../../errors/associations/MissingTh
 import ThroughAssociationConditionsIncompatibleWithThroughAssociationSource from '../../errors/associations/ThroughAssociationConditionsIncompatibleWithThroughAssociationSource.js'
 import CannotNegateSimilarityClause from '../../errors/CannotNegateSimilarityClause.js'
 import CannotPassUndefinedAsAValueToAWhereClause from '../../errors/CannotPassUndefinedAsAValueToAWhereClause.js'
+import DataTypeColumnTypeMismatch from '../../errors/db/DataTypeColumnTypeMismatch.js'
 import UnexpectedUndefined from '../../errors/UnexpectedUndefined.js'
 import CalendarDate from '../../helpers/CalendarDate.js'
 import camelize from '../../helpers/camelize.js'
@@ -558,21 +560,33 @@ export default class KyselyQueryDriver<DreamInstance extends Dream> extends Quer
 
     const sqlifiedAttributes = sqlAttributes(dream)
 
-    if (dream.isPersisted) {
-      const query = db
-        .updateTable(dream.table)
-        .set(sqlifiedAttributes as any)
-        .where(namespaceColumn(dream.primaryKey, dream.table), '=', dream.primaryKeyValue)
-      return await executeDatabaseQuery(
-        query.returning([...dream.columns()] as any),
-        'executeTakeFirstOrThrow'
-      )
-    } else {
-      const query = db
-        .insertInto(dream.table)
-        .values(sqlifiedAttributes as any)
-        .returning([...dream.columns()] as any)
-      return await executeDatabaseQuery(query, 'executeTakeFirstOrThrow')
+    try {
+      if (dream.isPersisted) {
+        const query = db
+          .updateTable(dream.table)
+          .set(sqlifiedAttributes as any)
+          .where(namespaceColumn(dream.primaryKey, dream.table), '=', dream.primaryKeyValue)
+        return await executeDatabaseQuery(
+          query.returning([...dream.columns()] as any),
+          'executeTakeFirstOrThrow'
+        )
+      } else {
+        const query = db
+          .insertInto(dream.table)
+          .values(sqlifiedAttributes as any)
+          .returning([...dream.columns()] as any)
+        return await executeDatabaseQuery(query, 'executeTakeFirstOrThrow')
+      }
+    } catch (error) {
+      switch (pgErrorType(error)) {
+        case 'INVALID_INPUT_SYNTAX':
+          throw new DataTypeColumnTypeMismatch({
+            dream,
+            error: error instanceof Error ? error : new Error('database column type error'),
+          })
+      }
+
+      throw error
     }
   }